+++ Src/MoneyManager.Business/Converter/BooleanToVisibilityConverter.cs
@@ -2,7 +2,7 @@
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Data;
 
-namespace MoneyManager.Converter
+namespace MoneyManager.Business.Converter
 {
     public class BooleanToVisibilityConverter : IValueConverter
     {
