+++ Src/MoneyManager.Business/Converter/RevertVisibilityConverter.cs
@@ -2,7 +2,7 @@
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Data;
 
-namespace MoneyManager.Converter
+namespace MoneyManager.Business.Converter
 {
     public class RevertVisibilityConverter : IValueConverter
     {
