+++ Src/MoneyManager/MoneyManager.Shared/UserControls/AccountListUserControl.xaml
@@ -4,14 +4,14 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:userControls="using:MoneyManager.UserControls"
-             xmlns:converter="using:MoneyManager.Converter"
+             xmlns:converter1="using:MoneyManager.Business.Converter"
              mc:Ignorable="d"
              d:DesignHeight="300"
              d:DesignWidth="400"
              DataContext="{Binding AccountListUserControlView, Source={StaticResource Locator}}">
 
     <UserControl.Resources>
-        <converter:AmountConverter x:Key="AmountConverter" />
+        <converter1:AmountConverter x:Key="AmountConverter" />
         <DataTemplate x:Key="AccountTemplate">
             <StackPanel Margin="0,0,0,7"
                         HorizontalAlignment="Stretch"
