--- Src/MoneyManager/MoneyManager.Shared/UserControls/AddTransactionUserControl.xaml
@@ -1,111 +1,0 @@
-﻿<UserControl
-    x:Class="MoneyManager.UserControls.AddTransactionUserControl"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:converter="using:MoneyManager.Converter"
-    xmlns:userControls="using:MoneyManager.UserControls"
-    mc:Ignorable="d"
-    d:DesignWidth="400" d:DesignHeight="800"
-    DataContext="{Binding AddTransactionView, Source={StaticResource Locator}}">
-
-    <UserControl.Resources>
-        <converter:AmountConverter x:Key="AmountConverter" />
-        <converter:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
-        <converter:BooleanToVisibilityConverter x:Key="BooleanVisibilityConverter" />
-        <DataTemplate x:Key="ComboBoxTemplate">
-            <StackPanel Margin="0,0,0,12" Orientation="Horizontal">
-                <TextBlock Text="{Binding Name}" FontSize="24" />
-                <TextBlock Text=" (" Style="{StaticResource ComboBoxItemSubText}" />
-                <TextBlock Text="{Binding CurrentBalance}" Style="{StaticResource ComboBoxItemSubText}"
-                           Margin="5,0,0,0" />
-                <TextBlock Text="{Binding Currency}" Style="{StaticResource ComboBoxItemSubText}" Margin="5,0,5,0" />
-                <TextBlock Text=")" Style="{StaticResource ComboBoxItemSubText}" />
-            </StackPanel>
-        </DataTemplate>
-    </UserControl.Resources>
-
-    <ScrollViewer>
-        <StackPanel Margin="12,0">
-            <ComboBox Header="charged account"
-                      x:Uid="ChargedAccount"
-                      ItemTemplate="{StaticResource ComboBoxTemplate}"
-                      ItemsSource="{Binding AllAccounts, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                      SelectedItem="{Binding SelectedTransaction.ChargedAccount, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}">
-            </ComboBox>
-            <ComboBox Header="target account" x:Uid="TargetAccount"
-                      ItemTemplate="{StaticResource ComboBoxTemplate}"
-                      ItemsSource="{Binding AllAccounts, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                      SelectedItem="{Binding SelectedTransaction.TargetAccount, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                      Visibility="{Binding IsTransfer, Converter={StaticResource BooleanVisibilityConverter}}" />
-            <Grid>
-                <TextBox
-                    Text="{Binding AmountString, Mode=TwoWay, 
-                    Converter={StaticResource AmountConverter}, UpdateSourceTrigger=PropertyChanged }"
-                    Header="amount"
-                    InputScope="Number"
-                    x:Name="TextBoxAmount"
-                    x:Uid="Amount"
-                    TextAlignment="Right"
-                    TextChanged="ReplaceSeparatorChar"
-                    GotFocus="RemoveZeroOnFocus"
-                    LostFocus="AddZeroIfEmpty" />
-
-                <Button MinWidth="40" MinHeight="55" Background="#FF231E1E"
-                        BorderThickness="0" Click="OpenSelectCurrencyDialog"
-                        Content="{Binding SelectedTransaction.Currency}"
-                        Margin="2,28,0,6" />
-            </Grid>
-            <StackPanel Orientation="Horizontal" HorizontalAlignment="Right"
-                        Visibility="{Binding SelectedTransaction.IsExchangeModeActive, Converter={StaticResource BooleanVisibilityConverter}}">
-
-                <TextBlock Text=" ( " Style="{StaticResource MenuItemSubSmallText}" />
-                <TextBlock Text="{Binding SelectedTransaction.ExchangeRatio}"
-                           Style="{StaticResource MenuItemSubSmallText}"
-                           TextAlignment="Right" />
-                <TextBlock Text=") " Style="{StaticResource MenuItemSubSmallText}" />
-
-                <TextBlock Text="{Binding SelectedTransaction.Amount}"
-                           Style="{StaticResource MenuItemSubSmallText}"
-                           Margin="5,0,0,0"
-                           TextAlignment="Right" />
-                <TextBlock Text="{Binding DefaultCurrency}"
-                           Margin="5,0,0,0"
-                           Style="{StaticResource MenuItemSubSmallText}" />
-            </StackPanel>
-            <Grid>
-                <userControls:SelectCategoryUserControl />
-            </Grid>
-            <DatePicker
-                Date="{Binding SelectedTransaction.Date, Converter={StaticResource DateTimeToDateTimeOffsetConverter}, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                Margin="0,12,0,0" Header="date"
-                x:Uid="DateHeader" />
-            <TextBox Header="Note"
-                     x:Uid="Notes"
-                     Text="{Binding SelectedTransaction.Note, Mode=TwoWay}"
-                     TextWrapping="Wrap" />
-            <ToggleSwitch Header="repeat" x:Uid="Repeat"
-                          VerticalAlignment="Stretch"
-                          IsOn="{Binding SelectedTransaction.IsRecurring, Mode=TwoWay}" />
-            <StackPanel
-                Visibility="{Binding SelectedTransaction.IsRecurring, Converter={StaticResource BooleanVisibilityConverter}}">
-                <ComboBox Header="Reccurence" x:Uid="RecurrenceHeader"
-                          SelectedIndex="{Binding Recurrence, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}">
-                    <ComboBoxItem Content="daily" x:Uid="DailyContent" />
-                    <ComboBoxItem Content="daily without weekend" x:Uid="DailyWithoutWeekendContent" />
-                    <ComboBoxItem Content="weekly" x:Uid="WeeklyContent" />
-                    <ComboBoxItem Content="monthly" x:Uid="MonthlyContent" />
-                    <ComboBoxItem Content="yearly" x:Uid="YearlyContent" />
-                </ComboBox>
-                <CheckBox Content="Endless" x:Uid="EndlessContent"
-                          IsChecked="{Binding IsEndless, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-                <DatePicker
-                    Date="{Binding EndDate, Converter={StaticResource DateTimeToDateTimeOffsetConverter}, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                    Margin="0,12,0,0" Header="enddate"
-                    Visibility="{Binding IsEndless, Converter={StaticResource BooleanVisibilityConverter}, ConverterParameter=revert}"
-                    x:Uid="EnddateHeader" />
-            </StackPanel>
-        </StackPanel>
-    </ScrollViewer>
-</UserControl>
