+++ Src/MoneyManager/MoneyManager.Shared/UserControls/BackupUserControl.xaml
@@ -3,9 +3,9 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:converter="using:MoneyManager.Converter"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:core="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:converter1="using:MoneyManager.Business.Converter"
     x:Class="MoneyManager.UserControls.BackupUserControl"
     mc:Ignorable="d"
     d:DesignHeight="300"
@@ -18,7 +18,7 @@
     </interactivity:Interaction.Behaviors>
 
     <UserControl.Resources>
-        <converter:BooleanToVisibilityConverter x:Key="BooleanVisibilityConverter" />
+        <converter1:BooleanToVisibilityConverter x:Key="BooleanVisibilityConverter" />
     </UserControl.Resources>
 
     <Grid>
