+++ Src/MoneyManager/MoneyManager.Shared/UserControls/BalanceUserControl.xaml
@@ -4,7 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:converter="using:MoneyManager.Converter"
+    xmlns:converter1="using:MoneyManager.Business.Converter"
     mc:Ignorable="d"
     d:DesignHeight="300"
     d:DesignWidth="400"
@@ -12,7 +12,7 @@
     Loaded="BalanceUserControl_OnLoaded">
 
     <UserControl.Resources>
-        <converter:AmountConverter x:Key="AmountConverter" />
+        <converter1:AmountConverter x:Key="AmountConverter" />
     </UserControl.Resources>
 
     <StackPanel Orientation="Horizontal">
