+++ Src/MoneyManager/MoneyManager.WindowsPhone/Dialogs/SelectStatisticDialog.xaml
@@ -4,7 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:converter="using:MoneyManager.Converter"
+    xmlns:converter1="using:MoneyManager.Business.Converter"
     mc:Ignorable="d"
     Title="SELECT STATISTIC"
     x:Uid="StatisticDialog"
@@ -14,7 +14,7 @@
     DataContext="{Binding StatisticView, Source={StaticResource Locator}}">
 
     <ContentDialog.Resources>
-        <converter:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
+        <converter1:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
     </ContentDialog.Resources>
